requirements.txt
fastapi
uvicorn
openai
pydanticgit 